backend/app/modules/stitcher/service.py
import ffmpeg
import os
import asyncio
import shutil

class StitcherService:
    def __init__(self):
        self.assets_dir = "app/modules/stitcher/assets"
        os.makedirs(self.assets_dir, exist_ok=True)
        # Ensure we have a mock background music file 
        # (For simulation, we'd need a real file, but code assumes it exists)
        self.bgm_path = os.path.join(self.assets_dir, "bgm.mp3")

    async def stitch_video(self, video_url: str, voice_path: str) -> str:
        """
        Agentic Stitching Logic:
        1. Input: Generated Video Loop (simulated via URL or local path) + Voiceover.
        2. Input: Background Music (local asset).
        3. Process:
           - Loop video to cover VO duration.
           - Mix VO and BGM.
           - Apply Sidechain Compression: BGM ducks when VO speaks.
        """
        print(f"Stitcher Engine: Processing {video_url} + {voice_path}...")
        
        # SIMULATION ONLY: Return the input video as we don't have real media files generated yet.
        # In a real run, we would download the video_url to a temp path.
        if "http" in video_url:
             print("Stitcher: Running in Simulation Mode (skipping actual ffmpeg verify for now)")
             await asyncio.sleep(2.5)
             return video_url

        # REAL LOGIC DRAFT (for when local files exist)
        # output_path = "output_render.mp4"
        # try:
        #     # 1. Inputs
        #     video = ffmpeg.input(video_url, stream_loop=-1) # Loop indefinitely
        #     voice = ffmpeg.input(voice_path)
        #     bgm = ffmpeg.input(self.bgm_path, stream_loop=-1) # Loop bgm
        #
        #     # 2. Audio Ducking Graph
        #     # "asplit" voice to use it as both audible track and sidechain trigger
        #     # "sidechaincompress" on bgm using voice as trigger
        #     # Arguments: threshold, ratio, attack, release
        #     ducked_bgm = ffmpeg.filter(
        #         [bgm, voice], 
        #         "sidechaincompress", 
        #         threshold=0.1, 
        #         ratio=4, 
        #         attack=50, 
        #         release=200
        #     )
        #     
        #     # 3. Mix Audio (Voice + Ducked BGM)
        #     final_audio = ffmpeg.filter([ducked_bgm, voice], "amix", inputs=2)
        #
        #     # 4. Trim Video to Voice Duration
        #     # We need to probe voice duration first to set -t for output
        #     probe = ffmpeg.probe(voice_path)
        #     duration = float(probe['format']['duration'])
        #
        #     # 5. Pipeline
        #     pipeline = (
        #         ffmpeg
        #         .output(video, final_audio, output_path, t=duration + 1, vcodec='libx264', acodec='aac')
        #         .overwrite_output()
        #     )
        #     pipeline.run()
        #     return output_path
        # except Exception as e:
        #     print(f"FFmpeg Error: {e}")
        #     return video_url

        return video_url
