backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
httpx
python-multipart
python-dotenv
openai
ffmpeg-python
